shared/pyproject.toml
[project]
name = "aoc_utils"
version = "0.1.0"
description = "Utilities for Advent of Code challenges"
authors = [{name = "Mosqueteiro"}]
requires-python = ">= 3.11"
dependencies = [
    "requests>=2.31.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
aoc_utils = { path = ".", editable = true }

[tool.pixi.dependencies]
python-dotenv = ">=1.2.1,<2"

[tool.pixi.tasks]
setup = "python -m aoc_utils.setup"
